Make salinity editable: Inconsistencies with manual dives
This solves some inconsistencies when setting the salinity of
a manual dive. This whole PR does not douch either
dive->salinity or dive->dc.salinity, even for manual dives. In theory
one should probably set dive->salinity for manually-entered dives,
but it does not at the moment.

In addition, I tracked a comment or two of @bstoeger that I did not
correct. I corrected those.

+++ commands/command_edit.cpp
@@ -358,7 +358,7 @@ DiveField EditWaterTemp::fieldId() const
 // ***** Water Type *****
 void EditWaterTypeUser::set(struct dive *d, int value) const
 {
-	d->user_salinity = value > 0 ? (uint32_t)value : 0u;
+	d->user_salinity = value > 0 ? value : 0u;
 }
 
 int EditWaterTypeUser::data(struct dive *d) const
